+++ financeager/cli.py
@@ -68,14 +68,6 @@ def main():
     sys.exit(exit_code)
 
 
-def deprecated_main():
-    logger.warning(
-        "The 'financeager' main command will be renamed to 'fina' in an "
-        "upcoming release.\nUse the 'fina' command to avoid this warning."
-    )
-    main()
-
-
 def run(command, configuration, plugins=None, verbose=False, sinks=None, **params):
     """Run 'command' request using additional 'params'.